Refactor form pattern
