+++ src/main/java/com/example/studentservice/form/ClassQuery.java
@@ -1,0 +1,9 @@
+package com.example.studentservice.form;
+
+import lombok.Data;
+
+@Data
+public class ClassQuery {
+    String className;
+    String classSerial;
+}
