+++ src/main/java/com/example/studentservice/form/ScoreQuery.java
@@ -1,0 +1,12 @@
+package com.example.studentservice.form;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ScoreQuery {
+    String studentName;
+    Double min;
+    Double max;
+}
