--- src/main/java/com/example/studentservice/query/GradeQuery.java
@@ -1,10 +1,0 @@
-package com.example.studentservice.query;
-
-import lombok.Data;
-
-@Data
-public class GradeQuery {
-    Double min;
-    Double max;
-    String studentName;
-}
