+++ src/main/resources/mappers/ClassInfoMapper.xml
@@ -16,8 +16,15 @@
         </association>
     </resultMap>
 
-    <select id="findLast" resultType="string">
+    <sql id="otherMax">
+        SELECT c_serial FROM classes WHERE c_id = (SELECT MAX(c_id) FROM classes);
+    </sql>
+
+    <sql id="max">
         SELECT c_serial FROM classes ORDER BY c_id DESC LIMIT 1;
+    </sql>
+    <select id="findLast" resultType="string">
+        <include refid="otherMax"/>
     </select>
 
     <select id="findAll" resultMap="classResult">
