+++ src/main/resources/mappers/GradeMapper.xml
@@ -72,7 +72,7 @@
 
     <select id="findAllByClassId" parameterType="int" resultMap="gradeResult">
         SELECT g_id, g_language, g_math, g_english, g_status, g_creator, g_creation_time, g_last_editor, g_last_time
-        FROM grades g left outer join students s on g.s_id = s.s_id
+        FROM grades g inner join students s on g.s_id = s.s_id
         WHERE s.c_id = #{classId};
     </select>
 
@@ -96,7 +96,8 @@
     </select>
 
     <select id="findByStudentId" parameterType="int" resultMap="gradeResult">
-       SELECT g_id, g_language, g_math, g_english, g_status, g_creator, g_creation_time, g_last_editor, g_last_time FROM grades WHERE s_id = #{id};
+       SELECT g_id, g_language, g_math, g_english, g_status, g_creator, g_creation_time, g_last_editor, g_last_time
+       FROM grades WHERE s_id = #{id};
     </select>
 
     <select id="findByStudentSerialAndName" parameterType="Student" resultMap="gradeResultWithStudent">
